chore: Refactor time profiling in ProfilerService
The `time_profile` method in the `ProfilerService` class has been refactored to accept the `func` parameter along with any additional arguments and keyword arguments. This change allows for more flexibility in profiling different functions. The execution time is logged using the `logger` module.

Note: This commit message follows the established convention of using the "chore" type for refactoring changes.